Fix: RGB type Int -> Float 으로 변경

+++ BanGiDa/Extension/UIColor+Extension.swift
@@ -22,7 +22,7 @@ extension UIColor {
     static let tabBarColor = UIColor(named: "TabBarColor")
     static let darkPink = UIColor(red: 228/255, green: 167/255, blue: 180/255, alpha: 1)
     
-    convenience init(r: Int, g: Int, b: Int, alpha: CGFloat = 1.0) {
+    convenience init(r: CGFloat, g: CGFloat, b: CGFloat, alpha: CGFloat = 1.0) {
         self.init(
             red: CGFloat(r / 255),
             green: CGFloat(g / 255),
